Server/requirements.txt
aiofiles==24.1.0
aiolimiter==1.2.1
annotated-types==0.7.0
anyio==4.9.0
anytree==2.13.0
asttokens==2.4.1
asyncio-atexit==1.0.1
autogen-agentchat==0.5.4
autogen-core==0.5.4
autogen-ext==0.5.4
autograd==1.7.0
azure-ai-inference==1.0.0b9
azure-ai-projects==1.0.0b10
azure-common==1.1.28
azure-core==1.33.0
azure-cosmos==4.9.0
azure-identity==1.21.0
azure-search-documents==11.5.2
azure-storage-blob==12.25.1
beartype==0.18.5
beautifulsoup4==4.13.3
blis==1.2.1
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
cymem==2.0.11
Deprecated==1.2.18
deprecation==2.1.0
devtools==0.12.2
distro==1.9.0
docker==7.1.0
environs==11.2.1
executing==2.2.0
fnllm==0.2.9
fonttools==4.56.0
future==1.0.0
gensim==4.3.3
graphrag==2.1.0
graspologic==3.4.1
graspologic-native==1.2.5
grpcio==1.70.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
hyppo==0.4.0
idna==3.10
importlib_metadata==8.6.1
isodate==0.7.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
json-schema-to-pydantic==0.2.2
json_repair==0.30.3
jsonref==1.1.0
kiwisolver==1.4.8
lancedb==0.17.0
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.44.0
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mcp==1.6.0
mcp-server-fetch==2025.1.17
mdurl==0.1.2
msal==1.32.0
msal-extensions==1.3.1
murmurhash==1.0.12
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
openai==1.69.0
opentelemetry-api==1.31.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
POT==0.9.5
preshed==3.0.9
Protego==0.4.0
protobuf==5.29.4
pyarrow==15.0.2
pycparser==2.22
pydantic==2.11.1
pydantic-settings==2.8.1
pydantic_core==2.33.0
Pygments==2.19.1
PyJWT==2.10.1
pylance==0.20.0
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
readabilipy==0.3.0
regex==2024.11.6
remotior_sensus==0.4.4
requests==2.32.3
rich==13.9.4
scikit-learn==1.6.1
scipy==1.12.0
seaborn==0.13.2
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
sse-starlette==2.2.1
starlette==0.46.1
statsmodels==0.14.4
tenacity==9.1.2
textblob==0.18.0.post0
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.8.0
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
umap-learn==0.5.7
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
weasel==0.4.1
webencodings==0.5.1
wheel==0.45.1
wrapt==1.17.2
zipp==3.21.0
